+++ vite.config.standalone.js
@@ -2,9 +2,9 @@ export default {
   base: "./",
   build: {
     lib: {
-      entry: "src/standalone.js",
+      entry: "src/index.js",
       formats: ["umd"],
-      name: "vtkWASM",
+      name: "vtkwasm",
       fileName: "vtk",
     },
     assetsDir: ".",